shuffling Library files around. Adding library page title to all library subpages
