+++ enterprise/frontend/src/metabase-enterprise/data-studio/library/LibraryPage/RootSnippetsCollectionMenu.unit.spec.tsx
