+++ enterprise/frontend/src/metabase-enterprise/data-studio/library/LibraryPage/index.ts
@@ -1,0 +1,1 @@
+export { LibraryPage } from "./LibraryPage";
