+++ enterprise/frontend/src/metabase-enterprise/data-studio/library/routes.tsx
@@ -2,6 +2,7 @@ import { IndexRoute, Route } from "react-router";
 
 import { getDataStudioSegmentRoutes } from "metabase/data-studio/segments/routes";
 
+import { LibraryPage } from "./LibraryPage";
 import { LibrarySectionLayout } from "./LibrarySectionLayout";
 import { getDataStudioMetricRoutes } from "./metrics/routes";
 import { getDataStudioSnippetRoutes } from "./snippets/routes";
@@ -9,8 +10,8 @@ import { getDataStudioTableRoutes } from "./tables/routes";
 
 export const getDataStudioLibraryRoutes = () => {
   return (
-    <Route path="library">
-      <IndexRoute component={LibrarySectionLayout} />
+    <Route path="library" component={LibrarySectionLayout}>
+      <IndexRoute component={LibraryPage} />
       {getDataStudioTableRoutes()}
       {getDataStudioMetricRoutes()}
       {getDataStudioSegmentRoutes()}
